Create a shadow root

A shadow DOM is a self-contained DOM tree, contained within another DOM tree. A shadow root is an element that contains a DOM subtree. This class represents a shadow root object, allowing you to select elements within the shadow root.

Value

A ShadowRoot object.

A WebElement object.

A list of WebElement objects.

A list, which can then be converted to JSON using jsonlite::toJSON().

Public fields

id

The id of the shadow root.

Methods

Public methods

ShadowRoot$new()

ShadowRoot$find_element()

ShadowRoot$find_elements()

ShadowRoot$toJSON()

ShadowRoot$clone()

Method new()

Initialize a new ShadowRoot object. This should not be called manually: instead use WebElement$shadow_root(), or SeleniumSession$create_shadow_root().

Usage

ShadowRoot$new(session_id, req, verbose, id)

Arguments

session_id

The id of the session.

req, verbose

Private fields of a SeleniumSession object.

id

The id of the shadow root.

Method find_element()

Find an element in the shadow root.

Usage

ShadowRoot$find_element(
  using = c("css selector", "xpath", "tag name", "link text", "partial link text"),
  value
)

Arguments

using

The type of selector to use.

value

The value of the selector: a string.

Method find_elements()

Find all elements in a shadow root matching a selector.

Usage

ShadowRoot$find_elements(
  using = c("css selector", "xpath", "tag name", "link text", "partial link text"),
  value
)

Arguments

using

The type of selector to use.

value

The value of the selector: a string.

Method toJSON()

Convert an element to JSON. This is used by SeleniumSession$execute_script().

Usage

ShadowRoot$toJSON()

Method clone()

The objects of this class are cloneable with this method.

Usage

ShadowRoot$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) { # selenium_server_available()
session <- SeleniumSession$new()

# Let's create our own Shadow Root using JavaScript
session$execute_script("
  const div = document.createElement('div');
  document.body.appendChild(div);
  div.attachShadow({mode: 'open'});
")

element <- session$find_element(using = "css selector", value = "div")

element$shadow_root()

session$close()
}
if (FALSE) { # selenium_server_available()
session <- SeleniumSession$new()

# Let's create our own Shadow Root using JavaScript
session$execute_script("
  const div = document.createElement('div');
  document.body.appendChild(div);
  const shadowRoot = div.attachShadow({mode: 'open'});
  const span = document.createElement('span');
  span.textContent = 'Hello';
  shadowRoot.appendChild(span);
")

element <- session$find_element(using = "css selector", value = "div")

shadow_root <- element$shadow_root()

shadow_root$find_element(using = "css selector", value = "span")

session$close()
}
if (FALSE) { # selenium_server_available()
session <- SeleniumSession$new()

# Let's create our own Shadow Root using JavaScript
session$execute_script("
  const div = document.createElement('div');
  document.body.appendChild(div);
  const shadowRoot = div.attachShadow({mode: 'open'});
  const span = document.createElement('span');
  span.textContent = 'Hello';
  shadowRoot.appendChild(span);
  const p = document.createElement('p');
  p.textContent = 'Me too!';
  shadowRoot.appendChild(p);
")

element <- session$find_element(using = "css selector", value = "div")

shadow_root <- element$shadow_root()

shadow_root$find_elements(using = "css selector", value = "*")

session$close()
}
if (FALSE) { # selenium_server_available()
session <- SeleniumSession$new()

# Let's create our own Shadow Root using JavaScript
session$execute_script("
  const div = document.createElement('div');
  document.body.appendChild(div);
  div.attachShadow({mode: 'open'});
")

element <- session$find_element(using = "css selector", value = "div")

shadow_root <- element$shadow_root()

result <- shadow_root$toJSON()

result

jsonlite::toJSON(result, auto_unbox = TRUE)

session$close()
}