Create a live element

This class represents a single element on the page. It is created using an existing SeleniumSession instance.

Value

A WebElement object.

A ShadowRoot object.

A WebElement object.

A list of WebElement objects.

A boolean value: TRUE or FALSE.

The value of the attribute: a string.

The value of the property: a string.

The value of the CSS property: a string.

The text content of the element: a string.

The tag name of the element: a string.

A list containing the following elements:

x: The x-coordinate of the element.

y: The y-coordinate of the element.

width: The width of the element in pixels.

height: The height of the element in pixels.

A boolean value: TRUE or FALSE.

A string.

A string.

The element, invisibly.

The element, invisibly.

The element, invisibly.

The base64-encoded PNG screenshot, as a string.

A list, which can then be converted to JSON using jsonlite::toJSON().

Public fields

id

The id of the element, used to uniquely identify it on the page.

Methods

Public methods

WebElement$new()

WebElement$shadow_root()

WebElement$find_element()

WebElement$find_elements()

WebElement$is_selected()

WebElement$get_attribute()

WebElement$get_property()

WebElement$get_css_value()

WebElement$get_text()

WebElement$get_tag_name()

WebElement$get_rect()

WebElement$is_enabled()

WebElement$computed_role()

WebElement$computed_label()

WebElement$click()

WebElement$clear()

WebElement$send_keys()

WebElement$screenshot()

WebElement$toJSON()

WebElement$clone()

Method new()

Initialize a WebElement object. This should not be called manually: instead use SeleniumSession$create_webelement() if you have an element id. To find elements on the page, use SeleniumSession$find_element() and SeleniumSession$find_elements().

Usage

WebElement$new(session_id, req, verbose, id)

Arguments

session_id

The id of the session that the element belongs to.

req, verbose

Private fields of a SeleniumSession object.

id

The element id.

Method shadow_root()

A shadow DOM is a self-contained DOM tree, contained within another DOM tree. A shadow root is an element that contains a DOM subtree. This method gets the shadow root property of an element.

Usage

WebElement$shadow_root()

Method find_element()

Find the first element matching a selector, relative to the current element.

Usage

WebElement$find_element(
  using = c("css selector", "xpath", "tag name", "link text", "partial link text"),
  value
)

Arguments

using

The type of selector to use.

value

The value of the selector: a string.

Method find_elements()

Find all elements matching a selector, relative to the current element.

Usage

WebElement$find_elements(
  using = c("css selector", "xpath", "tag name", "link text", "partial link text"),
  value
)

Arguments

using

The type of selector to use.

value

The value of the selector: a string.

Method is_selected()

Check if an element is currently selected.

Usage

WebElement$is_selected()

Method get_attribute()

Get an attribute from an element.

Usage

WebElement$get_attribute(name)

Arguments

name

The name of the attribute.

Method get_property()

Get a property from an element. Properties are similar to attributes, but represent the HTML source code of the page, rather than the current state of the DOM.

Usage

WebElement$get_property(name)

Arguments

name

The name of the property.

Method get_css_value()

Get the computed value of a CSS property.

Usage

WebElement$get_css_value(name)

Arguments

name

The name of the CSS property.

Method get_text()

Get the text content of an element.

Usage

WebElement$get_text()

Method get_tag_name()

Get the tag name of an element.

Usage

WebElement$get_tag_name()

Method get_rect()

Get the dimensions and coordinates of an element.

Usage

WebElement$get_rect()

Method is_enabled()

Check if an element is currently enabled.

Usage

WebElement$is_enabled()

Method computed_role()

Get the computed role of an element. The role of an element is usually "generic", but is often used when an elements tag name differs from its purpose. For example, a link that is "button-like" in nature may have a "button" role.

Usage

WebElement$computed_role()

Method computed_label()

Get the computed label of an element (i.e. The text of the label element that points to the current element).

Usage

WebElement$computed_label()

Method click()

Click on an element.

Usage

WebElement$click()

Method clear()

Clear the contents of a text input element.

Usage

WebElement$clear()

Method send_keys()

Send keys to an element.

Usage

WebElement$send_keys(...)

Arguments

...

The keys to send (strings). Use keys for special keys, and use key_chord() to send multiple keys at once.

Method screenshot()

Take a screenshot of an element.

Usage

WebElement$screenshot()

Method toJSON()

Convert an element to JSON. This is used by SeleniumSession$execute_script().

Usage

WebElement$toJSON()

Method clone()

The objects of this class are cloneable with this method.

Usage

WebElement$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.