Download and start the Selenium server.

Downloads the latest release of Selenium Server, and then runs it as a background process. You must have Java installed for this command to work.

Usage

selenium_server(
  version = "latest",
  selenium_manager = TRUE,
  interactive = TRUE,
  echo_cmd = FALSE,
  stdout = NULL,
  stderr = NULL,
  extra_args = c()
)

Arguments

version

The version of Selenium Server to download and run. By default, the latest major or minor release is used.

selenium_manager

Whether to enable Selenium Manager, which will automatically download any missing drivers. Defaults to TRUE.

interactive

By default, if you don't have a version downloaded, you will be prompted to confirm that you want to download it, and the function will error if rlang::is_interactive() returns FALSE. To allow this function to work in a non-interactive setting, set this to FALSE.

echo_cmd, stdout, stderr

Passed into processx::process$new().

extra_args

A character vector of extra arguments to pass into the Selenium Server call.

Value

A processx::process object. Call <process>$kill() to stop the server.

See also

Examples

if (FALSE) { # rlang::is_interactive()
if (FALSE) {
server <- selenium_server()

server$kill()
}
}